requirements.txt
pandas
numpy
joblib
scikit-learn==1.0.2
Flask
tensorflow-cpu
gunicorn